Image Processing Module
Ideas in this can be further tweaked or extended for the need of the algorithm. Currently, it will go into a specified path, and turn all images in that path to 75x75 greyscale images.

Focus on finding a good dataset to work with, processing those images, and applying Scikit algorithms next iteration.