Removed parts of the FillerStrategy class, Col drop update
Since we weren't really using the column types from get_composition, I
did away with using it entirely. I've left get_composition in
dataset_insight in case we need it at  some point, but since getting the
dtypes is so trivial, I did away with it.

Also updated how the column drop works in situations where the feature
needs to be dropped. Previously the column dropped would be filled with
"None", now it is irradicated entirely.